Rewrote the whole thing to make the process more sleeker and easy

+++ autometa.py
@@ -1,10 +1,11 @@
 import os
 import requests
 import eyed3
+from tabulate import tabulate
 
 print('\n','-'*19,'AUTOMETADATA','-'*19,'\n')
 
-location = input('[Enter location / name of music file(s) (Leave blank if current directory)] ')
+location = input('[Enter location / name of music file(s) (Leave blank if current directory)] ') + '/'
 
 # if blank
 if location.strip() == '' : location = os.getcwd()
@@ -19,20 +20,6 @@ elif os.path.isfile(location): files = [location]
 else:
     print('INVALID. Please check again.')
     exit()
-
-
-try:
-    download_method = int(input('Choose download method -\n\
-[1] Fully automatically (won\'t be that accurate) \n\
-[2] Only ask if there is trouble finding the song \n\
-[3] Ask for confirmation on every song \n\
-: '))
-
-# if no input given
-except ValueError:
-    print('Not a valid answer. Please try again.')
-    exit()
-
 
 print('\n','-'*50,'\n')
 
@@ -56,7 +43,6 @@ def convert_file_name_to_searchable_term(file):
 with open('api_key') as f:
     api_key = f.readline()
 
-print(api_key)
      
 def search_genius(search_term):
     url = f'https://api.genius.com/search?q={search_term}&page=1&per_page=3&access_token={api_key}'
@@ -70,23 +56,25 @@ def search_genius(search_term):
         exit()
 
 
-def get_metadata_from_shazam():
-    responses = shazam_response['response']['hits']
-    num_of_responses = len(responses)
+def get_metadata_from_response(response):
+    hits = response['response']['hits']
+    num_of_hits = len(hits)
 
     song_title_list = []
     artist_list = []
     album_art_list = []
 
-    for i in range(num_of_responses):
-        song_title_list.append(responses[i]['result']['title_with_featured'])
-        artist_list.append(responses[i]['result']['primary_artist']['name'])
-        album_art_list.append(responses[i]['result']['song_art_image_url'])
+    for i in range(num_of_hits):
+        song_title_list.append(hits[i]['result']['title_with_featured'])
+        artist_list.append(hits[i]['result']['primary_artist']['name'])
+        album_art_list.append(hits[i]['result']['song_art_image_url'])
 
     return [song_title_list , artist_list , album_art_list]
 
 
-def apply_metadata():
+def apply_metadata(file_name):
+    print('Applying Metadata for',file_name)
+
     #apply song title
     audio_file.tag.title = song_title
     print('- Applied Title -',song_title)
@@ -98,16 +86,17 @@ def apply_metadata():
     #apply album name
     audio_file.tag.album = artist #done just so album art displays correctly on mobile
 
-    #apply album art
-    image = requests.get(album_art)
+    #apply album art if it is not '-' (which is kept when metadata is not received from genius even after implicit search)
+    if album_art != '-':
+        image = requests.get(album_art)
 
-    with open('album_art.jpg','wb') as img_file:
-        img_file.write(image.content)
+        with open('album_art.jpg','wb') as img_file:
+            img_file.write(image.content)
 
-    audio_file.tag.images.set(3, open('album_art.jpg','rb').read(), 'image/jpeg')
-    os.remove('album_art.jpg')
+        audio_file.tag.images.set(3, open('album_art.jpg','rb').read(), 'image/jpeg')
+        os.remove('album_art.jpg')
 
-    print('- Applied Album Art')
+        print('- Applied Album Art')
 
 
 # makes the file name artist + song name
@@ -119,86 +108,158 @@ def change_file_name():
     print('- Changed file name')
 
 
-def ask_for_name_of_song_and_artist():
-    name = input('[Song Name] ')
-    artist = input('[Artist Name] ')
-    print()
-    return [name,artist]
+metadata = { 'available': [] , 'in_use': [] }
+
+files_without_metadata_received = []
+files_without_proper_metadata_shown = []
+
+# Searches genius for metadata based on file name
+# Adds it to the dict metadata as available (complete metadata received) and as in_use (the first hits of metadata received)
+for i in range(len(files)):
+    file = files[i]
+    print(f'[{i+1}/{len(files)}] {file}')
+    
+    # searches and gets the response from genius
+    search_term = convert_file_name_to_searchable_term(file)
+    genius_response = search_genius(search_term)
+
+    # decodes the response into lists of song_title artist , and album art
+    song_title_list , artist_list , album_art_list = get_metadata_from_response(genius_response)
+
+    # adds all metadata to available key of metadata
+    metadata['available'].extend([[file,song_title_list,artist_list,album_art_list]])
+
+    if len(song_title_list) == 0 :
+        files_without_metadata_received.append(file) # This list is mode so that in the end we can ask input for implicit search
+        song_title_list.append('-')
+        artist_list.append('-')
+        album_art_list.append('-')
+         
+
+    # adds only the 1st metadata of each song to in_use key of metadata
+    metadata['in_use'].extend([[ #remove i+1
+        file,
+        song_title_list[0],
+        artist_list[0],
+        album_art_list[0]
+    ]])
+     
+    
+
+# displays the table for confirmation
+print('\n',tabulate(
+        [metadata['in_use'][i][:-1] for i in range(len(metadata['in_use']))],
+        headers=['ID','File name','Song Title','Artist'],
+        showindex=True
+    ))
 
 
-for index,file in enumerate(files):
-    print('\n','-'*50,'\n')
+ids_of_wrong_metadata = list(map(int,input('\n[Enter ID of music files with wrong metadata (eg: \' 0 3 15 \' ) (Leave blank if none) (Ignore ones with - in columns) ] ').split()))
 
-    print(f'[{index + 1}] {file}\n')
+ids_of_wrong_metadata_in_range = []
+ids_of_wrong_metadata_out_of_range = []
 
-    file_name = convert_file_name_to_searchable_term(file)
+for id in ids_of_wrong_metadata:
+    if id-1 < len(files) :
+        ids_of_wrong_metadata_in_range.append(id)
+    else:
+        ids_of_wrong_metadata_out_of_range.append(id)
 
-    shazam_response = search_genius(file_name)
+# gets the file names of those without proper metadata using index from ids_of_wrong_metadata_in_range
+files_without_proper_metadata_shown = [ [i,metadata['available'][i]] for i in ids_of_wrong_metadata_in_range ]
 
-    was_shazam_searched_with_explicit_name = False
+# For checking if other metadata responses is correct
+for file_info in files_without_proper_metadata_shown:
+    index_of_file = file_info[0] 
+    file = file_info[1][0]
+    
+    for i in range(1,len(metadata['available'][index_of_file][1])): # starts from the 2nd result because first one was displayed
 
-    index_of_response_used = 0
+        song_title = metadata['available'][index_of_file][1][i]
+        artist = metadata['available'][index_of_file][2][i]
+        album_art = metadata['available'][index_of_file][3][i]
 
-    while True:
-        try:
-            # so that it asks user for song title and artist
-            if index_of_response_used == 3: raise KeyError()
+        print(f'\nIs {file} : \n- {song_title}\n- by {artist}')
 
-            song_title_list , artist_list , album_art_list = get_metadata_from_shazam()
-
-            if len(song_title_list) == 0 : raise KeyError() # no result
-
-            if download_method == 3:
-                is_it_correct_response = input(f'[Is this {song_title_list[index_of_response_used]} by {artist_list[index_of_response_used]} ? ( y / n ) ] ') in ('y','')
-
-                if not is_it_correct_response : 
-                    index_of_response_used += 1 
-                    continue
-
-            
-            # get the audio file
-            path_to_file = f'{location}/{file}'
-            audio_file = eyed3.load(path_to_file)
-
-            # Get the required metadata from list
-            song_title = song_title_list[index_of_response_used]
-            artist = artist_list[index_of_response_used]
-            album_art = album_art_list[index_of_response_used]
-
-            apply_metadata()
-            
-            audio_file.tag.save(version=eyed3.id3.ID3_V2_3)
-            change_file_name() 
+        is_correct = input('[y / n] ') in ('y','')
         
-            print('\n','-'*50,'\n')
+        print()
 
+        if is_correct:
+            print('-'*50)
+            metadata['in_use'][index_of_file] = [file,song_title,artist,album_art]
             break
 
-        # arises when Shazam doesnt give adequate info
-        except (KeyError , TypeError): 
-            print('\nWe were unable to find result for',file)
-            
-            if not was_shazam_searched_with_explicit_name and download_method != 1:
-                print('Please give the name of song and artist for further checking. (Leave blank if you are not sure.)\n')
-                
-                song_name_input_given , artist_input_given = ask_for_name_of_song_and_artist()
+    else:
+        # Adds it to list of files without metadata since none of the 3 received was correct
+        # doesnt add if this is one without metadata from the start
+        if file not in files_without_metadata_received : 
+            files_without_metadata_received.append(file)
 
-                new_search_term = f'{song_name_input_given} {artist_input_given}'
-                shazam_response = search_genius(new_search_term)
-                
-                was_shazam_searched_with_explicit_name = True
 
-                index_of_response_used = 0
+# removes info of files for which metadata was not received
+for file in files_without_metadata_received:
+    for i in metadata['in_use']:
+        if file in i :
+            metadata['in_use'].remove(i)
 
-                continue # makes it try getting response from shazam 
-            
-            print('\n','-'*50,'\n')   
 
-            index_of_response_used = 0
+# Asks implicitly for song title and artist name, then searches genius
+for file in files_without_metadata_received:
+    print(f'\nFor {file} , What is the ')
+    
+    song_title_input = input('[Song Title] ')
+    artist_input = input('[Artist] ')
+
+    search_term = f'{song_title_input} {artist_input}'
+
+    # searches and gets the response from genius
+    search_term = convert_file_name_to_searchable_term(search_term)
+    genius_response = search_genius(search_term)
+
+    # decodes the response into lists of song_title artist , and album art
+    song_title_list , artist_list , album_art_list = get_metadata_from_response(genius_response)
+    
+    for i in range(len(song_title_list)):
+        song_title = song_title_list[i]
+        artist = artist_list[i]
+        album_art = album_art_list[i]
+
+        print(f'\nIs {file} : \n- {song_title}\n- by {artist}')
+        is_correct = input('[y / n] ') in ('y','')
+        print()
+        
+        if is_correct: 
             break
 
-        except Exception as err:
-            print('An error occured -',err)
-            print(file,'skipped.')
+    else:
+        # If none of these were correct 
+        print("Could not find proper metadata for this. Will be applying only title and artist.")
+        song_title = song_title_input.title()
+        artist = artist_input.title()
+        album_art = '-'
+    
+    metadata['in_use'].append([file,song_title,artist,album_art])
 
-            break
+print('-'*50)
+
+# applying the metadata to files
+
+for file_info in metadata['in_use']:
+    # get the audio file
+    path_to_file = f'{location}/{file_info[0]}'
+    audio_file = eyed3.load(path_to_file)
+
+    # Get the required metadata from list
+    song_title = file_info[1]
+    artist = file_info[2]
+    album_art = file_info[3]
+
+    apply_metadata(file_info[0])
+
+    audio_file.tag.save(version=eyed3.id3.ID3_V2_3)
+    change_file_name() 
+
+    print()
+
+print('-'*19,'END','-'*19)
